that's all.

+++ src/main/kotlin/dev/deepslate/fallacy/common/block/BlockWithHeat.kt
@@ -6,8 +6,8 @@ import net.minecraft.world.level.block.state.BlockState
 
 interface BlockWithHeat {
     //外延热量：方块对外部环境的影响热量，用于参与热力学引擎的计算影响环境温度
-    fun getEpitaxialHeat(state: BlockState, level: Level, pos: BlockPos): Float
+    fun getEpitaxialHeat(state: BlockState, level: Level, pos: BlockPos): Int
 
     //内秉热量: 方块自身蕴含的热量，用于方块功能
-    fun getIntrinsicHeat(state: BlockState, level: Level, pos: BlockPos): Float
+    fun getIntrinsicHeat(state: BlockState, level: Level, pos: BlockPos): Int
 }
